+++ src/main/java/ru/yandex/practicum/filmorate/controller/MpaController.java
@@ -25,7 +25,7 @@ public class MpaController {
     }
 
     @GetMapping("/{id}")
-    public Mpa findMpaById(@PathVariable("id") int id) {
+    public Mpa findMpaById(@PathVariable("id") Long id) {
         log.info("GET / '/mpa/{id}'");
         return mpaService.findMpaById(id);
     }
