+++ src/main/java/ru/yandex/practicum/filmorate/storage/LikeStorage.java
@@ -1,8 +1,8 @@
 package ru.yandex.practicum.filmorate.storage;
 
 public interface LikeStorage {
-    void addLike(int id, int userId);
+    void addLike(Long id, Long userId);
 
-    void removeLike(int id, int userId);
+    void removeLike(Long id, Long userId);
 }
